Centralize all URIs into SharedConstants.swift
Create new Shared/Utilities/SharedConstants.swift as single source of
truth for all URIs and paths across the project:

- Move skyLightFrameworkPath from Shims.swift to SharedConstants
- Move all app URLs (repository, issues, donate, spacing executable) from
  Constants.swift to SharedConstants
- Include Info.plist keys for URL configuration in SharedConstants
- Add private helper requiredInfoPlistURL() in SharedConstants

Update all usages to use SharedConstants directly:
- AboutSettingsPane.swift: SharedConstants.repositoryURL, issuesURL, donateURL
- MenuBarItemSpacingManager.swift: SharedConstants.menuBarItemSpacingExecutableURL
- Shims.swift: SharedConstants.skyLightFrameworkPath

Simplify Constants.swift to only contain bundle-specific constants
(version, build, copyright, bundleIdentifier, displayName).
